+++ include/pylon_camera/internal/pylon_camera.h
@@ -80,7 +80,7 @@ public:
     virtual bool setBinningY(const size_t& target_binning_y,
                              size_t& reached_binning_y);
 
-    virtual bool setImageEncoding(const std::string& target_ros_encoding);
+    virtual std::string setImageEncoding(const std::string& target_ros_encoding);
 
     virtual bool setExposure(const float& target_exposure, float& reached_exposure);